2000-08-28 10:35 UTC+0800 Ron Pinkas <ron@profit-master.com>    * source/compiler/harbour.c    * source/compiler/harbour.y      + Added support for DECLARE FunName(...) AS CLASS ClassName      + Added support for DECLARE FunName( ... @SomeFun() ... ) - Function Pointer declared parameter.
   * source/compiler/simplex.c
     * Corrected some compiler warnings.

   * source/compiler/harbour.slx
     * Reenabled support for ID_ON_HOLD.

   * include/hberror.h
     + Added: #define HB_COMP_WARN_RETURN_SUSPECT             28

   * source/compiler/hbgenerr.c
     + Added: "4Suspecious return type: \'%s\' expected: \'%s\'"

   * source/compiler/hbpcode.c
     + Completed support for Strong Typed variables as declared classes, as well as Adaptive Typed usage with declared classes.
     * Few other corrections and enhancements.

   * include/hbclass.ch
     + Added TClass declaration:
	DECLARE TClass ;
		New( cName AS STRING, OPTIONAL SuperParams ) AS CLASS TClass ;
		Create() AS OBJECT;
		Instance() AS OBJECT ;
		AddClsMthds( cName AS STRING, @Method(), nScope AS NUMERIC, n2 AS NUMERIC, n3 AS NUMERIC );
		AddMultiClsData( cType AS STRING, uVal, nScope AS NUMERIC, aDatas AS ARRAY OF STRING );
		AddMultiData( cType AS STRING, uVal, nScope AS NUMERIC, aDatas AS ARRAY OF STRING );
		AddMethod( cName AS STRING, @Method(), nScope AS NUMERIC );
		AddInLine( cName AS STRING, bBlock AS CODEBLOCK, nScope AS NUMERIC );
		AddVirtual( cName AS STRING )
     + Added declaration to s_oClass  AS CLASS TClass

   * tests/testdecl.prg
     * Minor refinement.

+++ harbour/source/compiler/simplex.c
@@ -1238,8 +1238,8 @@ YY_DECL
                   }
                #endif
 
-               i = aKeyNodes[ sToken[0] ].iMin;
-               iMax = aKeyNodes[ sToken[0] ].iMax + 1;
+               i = aKeyNodes[ (int) sToken[0] ].iMin;
+               iMax = aKeyNodes[ (int) sToken[0] ].iMax + 1;
                DEBUG_INFO( printf(  "Scanning Keys for Token: %s at Positions: %i-%i\n", (char*) sToken, i, iMax -1 ) );
 
                while ( i < iMax )
@@ -1298,8 +1298,8 @@ YY_DECL
                   }
                #endif
 
-               i = aWordNodes[ sToken[0] ].iMin;
-               iMax = aWordNodes[ sToken[0] ].iMax + 1;
+               i = aWordNodes[ (int) sToken[0] ].iMin;
+               iMax = aWordNodes[ (int) sToken[0] ].iMax + 1;
                DEBUG_INFO( printf(  "Scanning Words for Token: %s at Positions: %i-%i\n", (char*) sToken, i, iMax - 1 ) );
 
                while ( i < iMax )
